PathDict takes path_on_missing kwarg.  Fixes #2
This is a huge performance fix, since the previous code would create a PathDict
at every segment.  Each lookup would then go through the next PathDict's __getitem__
which involved a call to traverse, which did a bunch of work.  This makes the factory
configurable, and defaults to a dict.

As expected, this doesn't change the behavior at all, since it was the root PathDict
doing traversal, and the branch/leaf nodes don't need to know what's going on.