Split out context creation, layering responsibilities
This reduces the work of Context to (nearly) a single function.
While a class with one function and an __init__ is usually better
off as a function, ContextView requires a reference to the root
when including more contexts.  For now, readability is much improved.

The main win is getting rid of the special casing for the root,
and removing another function call for each __[get|set|del]item__.